Update: divide files
